% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{BB150803_IV_074}
\alias{BB150803_IV_074}
\alias{bam}
\title{Example BAM-file}
\format{A BAM file with aligned reads.}
\source{
A single cell from T-cell Acute Lymphoblastic Leukemia in mouse. The file has been downsampled to 20\% of the coverage to reduce the file size. For further details see TODO: insert publication (tumor ID T386).
}
\description{
A BAM-file for demonstration purposes of the functions in package \pkg{AneuFinder}.
}

