% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{hg19_diploid}
\alias{diploid}
\alias{hg19_diploid}
\title{Diploid BAM-file}
\format{A BED file with aligned reads.}
\source{
A single cell with WGSCS reads of a human brain cell. Unpublished data.
}
\description{
A BAM-file with reads from a diploid human brain cell for demonstration purposes of the functions in package \pkg{AneuFinder}.
}

