% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{metastasis_liver}
\alias{liver}
\alias{metastasis}
\alias{metastasis_liver}
\title{Liver metastasis results}
\format{Files with \code{aneuHMM} objects.}
\source{
Patient material from a liver metastasis of a 79 year old women. For details see TODO:insert publication.
}
\description{
Liver metastasis results in the form of Hidden Markov Models saved as RData, generated by the \pkg{AneuFinder} package.
}

