% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{primary_lung}
\alias{lung}
\alias{primary}
\alias{primary_lung}
\title{Primary lung cancer results}
\format{Files with \code{aneuHMM} objects.}
\source{
Patient material from a primary small cell lung cancer of a 79 year old women. For details see TODO:insert publication.
}
\description{
Primary lung cancer results in the form of Hidden Markov Models saved as RData, generated by the \pkg{AneuFinder} package.
}

