% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{KK150311_VI_07}
\alias{KK150311_VI_07}
\alias{bed}
\title{Example BED-file}
\format{A BED file with aligned reads.}
\source{
A single cell with Strand-seq reads. Unpublished data.
}
\description{
A BED-file for demonstration purposes of the functions in package \pkg{AneuFinder}.
}

